Extract enums to their own file in #enums
