+++ src/data/battle-anims.ts
@@ -3,10 +3,10 @@ import {
   AttackMove,
   BeakBlastHeaderAttr,
   DelayedAttackAttr,
-  MoveFlags,
   SelfStatusMove,
   allMoves,
 } from "./moves/move";
+import { MoveFlags } from "#enums/MoveFlags";
 import type Pokemon from "../field/pokemon";
 import {
   type nil,
