+++ src/data/moves/move.ts
@@ -120,89 +120,11 @@ import { LoadMoveAnimPhase } from "#app/phases/load-move-anim-phase";
 import { PokemonTransformPhase } from "#app/phases/pokemon-transform-phase";
 import { MoveAnimPhase } from "#app/phases/move-anim-phase";
 import { loggedInUser } from "#app/account";
-
-export enum MoveCategory {
-  PHYSICAL,
-  SPECIAL,
-  STATUS
-}
-
-export enum MoveTarget {
-  /** {@link https://bulbapedia.bulbagarden.net/wiki/Category:Moves_that_target_the_user Moves that target the User} */
-  USER,
-  OTHER,
-  ALL_OTHERS,
-  NEAR_OTHER,
-  /** {@link https://bulbapedia.bulbagarden.net/wiki/Category:Moves_that_target_all_adjacent_Pok%C3%A9mon Moves that target all adjacent Pokemon} */
-  ALL_NEAR_OTHERS,
-  NEAR_ENEMY,
-  /** {@link https://bulbapedia.bulbagarden.net/wiki/Category:Moves_that_target_all_adjacent_foes Moves that target all adjacent foes} */
-  ALL_NEAR_ENEMIES,
-  RANDOM_NEAR_ENEMY,
-  ALL_ENEMIES,
-  /** {@link https://bulbapedia.bulbagarden.net/wiki/Category:Counterattacks Counterattacks} */
-  ATTACKER,
-  /** {@link https://bulbapedia.bulbagarden.net/wiki/Category:Moves_that_target_one_adjacent_ally Moves that target one adjacent ally} */
-  NEAR_ALLY,
-  ALLY,
-  USER_OR_NEAR_ALLY,
-  USER_AND_ALLIES,
-  /** {@link https://bulbapedia.bulbagarden.net/wiki/Category:Moves_that_target_all_Pok%C3%A9mon Moves that target all Pokemon} */
-  ALL,
-  USER_SIDE,
-  /** {@link https://bulbapedia.bulbagarden.net/wiki/Category:Entry_hazard-creating_moves Entry hazard-creating moves} */
-  ENEMY_SIDE,
-  BOTH_SIDES,
-  PARTY,
-  CURSE
-}
-
-export enum MoveFlags {
-  NONE              = 0,
-  MAKES_CONTACT     = 1 << 0,
-  IGNORE_PROTECT    = 1 << 1,
-  /**
-   * Sound-based moves have the following effects:
-   * - Pokemon with the {@linkcode Abilities.SOUNDPROOF Soundproof Ability} are unaffected by other Pokemon's sound-based moves.
-   * - Pokemon affected by {@linkcode Moves.THROAT_CHOP Throat Chop} cannot use sound-based moves for two turns.
-   * - Sound-based moves used by a Pokemon with {@linkcode Abilities.LIQUID_VOICE Liquid Voice} become Water-type moves.
-   * - Sound-based moves used by a Pokemon with {@linkcode Abilities.PUNK_ROCK Punk Rock} are boosted by 30%. Pokemon with Punk Rock also take half damage from sound-based moves.
-   * - All sound-based moves (except Howl) can hit Pokemon behind an active {@linkcode Moves.SUBSTITUTE Substitute}.
-   *
-   * cf https://bulbapedia.bulbagarden.net/wiki/Sound-based_move
-   */
-  SOUND_BASED       = 1 << 2,
-  HIDE_USER         = 1 << 3,
-  HIDE_TARGET       = 1 << 4,
-  BITING_MOVE       = 1 << 5,
-  PULSE_MOVE        = 1 << 6,
-  PUNCHING_MOVE     = 1 << 7,
-  SLICING_MOVE      = 1 << 8,
-  /**
-   * Indicates a move should be affected by {@linkcode Abilities.RECKLESS}
-   * @see {@linkcode Move.recklessMove()}
-   */
-  RECKLESS_MOVE     = 1 << 9,
-  /** Indicates a move should be affected by {@linkcode Abilities.BULLETPROOF} */
-  BALLBOMB_MOVE     = 1 << 10,
-  /** Grass types and pokemon with {@linkcode Abilities.OVERCOAT} are immune to powder moves */
-  POWDER_MOVE       = 1 << 11,
-  /** Indicates a move should trigger {@linkcode Abilities.DANCER} */
-  DANCE_MOVE        = 1 << 12,
-  /** Indicates a move should trigger {@linkcode Abilities.WIND_RIDER} */
-  WIND_MOVE         = 1 << 13,
-  /** Indicates a move should trigger {@linkcode Abilities.TRIAGE} */
-  TRIAGE_MOVE       = 1 << 14,
-  IGNORE_ABILITIES  = 1 << 15,
-  /** Enables all hits of a multi-hit move to be accuracy checked individually */
-  CHECK_ALL_HITS    = 1 << 16,
-  /** Indicates a move is able to bypass its target's Substitute (if the target has one) */
-  IGNORE_SUBSTITUTE = 1 << 17,
-  /** Indicates a move is able to be redirected to allies in a double battle if the attacker faints */
-  REDIRECT_COUNTER  = 1 << 18,
-  /** Indicates a move is able to be reflected by {@linkcode Abilities.MAGIC_BOUNCE} and {@linkcode Moves.MAGIC_COAT} */
-  REFLECTABLE       = 1 << 19,
-}
+import { MoveCategory } from "#enums/MoveCategory";
+import { MoveTarget } from "#enums/MoveTarget";
+import { MoveFlags } from "#enums/MoveFlags";
+import { MoveEffectTrigger } from "#enums/MoveEffectTrigger";
+import { MultiHitType } from "#enums/MultiHitType";
 
 type MoveConditionFunc = (user: Pokemon, target: Pokemon, move: Move) => boolean;
 type UserMoveConditionFunc = (user: Pokemon, move: Move) => boolean;
@@ -1201,14 +1123,6 @@ export abstract class MoveAttr {
   }
 }
 
-export enum MoveEffectTrigger {
-  PRE_APPLY,
-  POST_APPLY,
-  HIT,
-  /** Triggers one time after all target effects have applied */
-  POST_TARGET,
-}
-
 interface MoveEffectAttrOptions {
   /**
    * Defines when this effect should trigger in the move's effect order
@@ -1918,14 +1832,6 @@ export class AddSubstituteAttr extends MoveEffectAttr {
       return i18next.t("moveTriggers:substituteNotEnoughHp");
     }
   }
-}
-
-export enum MultiHitType {
-  _2,
-  _2_TO_5,
-  _3,
-  _10,
-  BEAT_UP,
 }
 
 /**
