+++ src/enums/MoveCategory.ts
@@ -1,0 +1,5 @@
+export enum MoveCategory {
+  PHYSICAL,
+  SPECIAL,
+  STATUS
+}
