+++ src/enums/MoveEffectTrigger.ts
@@ -1,0 +1,7 @@
+export enum MoveEffectTrigger {
+  PRE_APPLY,
+  POST_APPLY,
+  HIT,
+  /** Triggers one time after all target effects have applied */
+  POST_TARGET
+}
