+++ src/enums/MultiHitType.ts
@@ -1,0 +1,7 @@
+export enum MultiHitType {
+  _2,
+  _2_TO_5,
+  _3,
+  _10,
+  BEAT_UP
+}
