+++ test/abilities/friend_guard.test.ts
@@ -6,7 +6,8 @@ import Phaser from "phaser";
 import { afterEach, beforeAll, beforeEach, describe, expect, it, vi } from "vitest";
 import { BattlerIndex } from "#app/battle";
 import { allAbilities } from "#app/data/ability";
-import { allMoves, MoveCategory } from "#app/data/moves/move";
+import { allMoves } from "#app/data/moves/move";
+import { MoveCategory } from "#enums/MoveCategory";
 
 describe("Moves - Friend Guard", () => {
   let phaserGame: Phaser.Game;
